add album view model
